# docs/api/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "ChallengeLevel": {
        "enum": ["easy", "medium", "hard"],
        "title": "ChallengeLevel",
        "type": "string"
      },
      "ChallengeRead": {
        "properties": {
          "challenge_number": {
            "title": "Challenge Number",
            "type": "integer"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "level": {
            "$ref": "#/components/schemas/ChallengeLevel"
          },
          "tag": {
            "$ref": "#/components/schemas/ChallengeTag"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "challenge_number",
          "tag",
          "level",
          "title",
          "content"
        ],
        "title": "ChallengeRead",
        "type": "object"
      },
      "ChallengeRunRequest": {
        "properties": {
          "prompt": {
            "maxLength": 4000,
            "minLength": 10,
            "title": "Prompt",
            "type": "string"
          },
          "publish_to_board": {
            "default": true,
            "title": "Publish To Board",
            "type": "boolean"
          }
        },
        "required": ["prompt"],
        "title": "ChallengeRunRequest",
        "type": "object"
      },
      "ChallengeRunResponse": {
        "properties": {
          "challenge": {
            "$ref": "#/components/schemas/ChallengeRead"
          },
          "output": {
            "title": "Output",
            "type": "string"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "raw": {
            "additionalProperties": true,
            "title": "Raw",
            "type": "object"
          },
          "share": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChallengeRunShareRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "usage": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Usage",
            "type": "object"
          }
        },
        "required": [
          "challenge",
          "prompt",
          "provider",
          "output",
          "usage",
          "raw",
          "share"
        ],
        "title": "ChallengeRunResponse",
        "type": "object"
      },
      "ChallengeRunShareRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_public": {
            "title": "Is Public",
            "type": "boolean"
          }
        },
        "required": ["id", "is_public", "created_at"],
        "title": "ChallengeRunShareRead",
        "type": "object"
      },
...
      "ChallengeTag": {
        "enum": ["ps", "img", "video"],
        "title": "ChallengeTag",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Prompteer API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/challenges": {
      "get": {
        "operationId": "list_challenges_api_v1_challenges_get",
        "parameters": [
          {
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ChallengeTag"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChallengeRead"
                  },
                  "title": "Response List Challenges Api V1 Challenges Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Challenges",
        "tags": ["challenges"]
      }
    },
    "/api/v1/challenges/{challenge_id}": {
      "get": {
        "operationId": "get_challenge_api_v1_challenges__challenge_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "challenge_id",
            "required": true,
            "schema": {
              "title": "Challenge Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Challenge",
        "tags": ["challenges"]
      }
    },
    "/api/v1/challenges/{challenge_id}/run": {
      "post": {
        "operationId": "run_challenge_prompt_api_v1_challenges__challenge_id__run_post",
        "parameters": [
          {
            "in": "path",
            "name": "challenge_id",
            "required": true,
            "schema": {
              "title": "Challenge Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChallengeRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Challenge Prompt",
        "tags": ["challenges"]
      }
    },
...
  }
}